simbarpmdashboard--main/advanced_ehs_simba/requirements.txt
# Advanced EHS Simba Drill Monitoring System
# Python 3.10+ Required

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
asyncpg>=0.29.0  # PostgreSQL async driver
alembic>=1.12.0  # Database migrations
supabase>=2.0.0  # Supabase Python SDK

# Data Processing & ML
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0

# IoT & Messaging
paho-mqtt>=1.6.1
aiomqtt>=2.0.0  # Async MQTT client
aiohttp>=3.9.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Notifications
httpx>=0.25.0  # Async HTTP client for webhooks
aiosmtplib>=3.0.0  # Async SMTP for email

# Monitoring & Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For FastAPI testing

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tenacity>=8.2.0  # Retry logic

# Arduino MVP
flask>=3.0.0
flask-cors>=4.0.0
pyserial>=3.5

